LPS-54821 Move Uuid change to 7_0_0 upgrade process
