+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_0.java
@@ -18,6 +18,7 @@ import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeAddress;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeAsset;
+import com.liferay.portal.upgrade.v7_0_0.UpgradeAssetPublisher;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeAssetTagsResourcePermission;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeBackgroundTask;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeCalEvent;
@@ -61,6 +62,7 @@ public class UpgradeProcess_7_0_0 extends UpgradeProcess {
 
 		upgrade(UpgradeAddress.class);
 		upgrade(UpgradeAsset.class);
+		upgrade(UpgradeAssetPublisher.class);
 		upgrade(UpgradeAssetTagsResourcePermission.class);
 		upgrade(UpgradeBackgroundTask.class);
 		upgrade(UpgradeCalEvent.class);
