+++ portal-impl/src/com/liferay/portal/upgrade/v6_0_0/UpgradeAssetPublisher.java
@@ -14,7 +14,6 @@
 
 package com.liferay.portal.upgrade.v6_0_0;
 
-import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -25,10 +24,6 @@ import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
 import com.liferay.portlet.PortletPreferencesFactoryUtil;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 
 import javax.portlet.PortletPreferences;
 
@@ -58,41 +53,6 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 		}
 
 		return assetEntryXmls;
-	}
-
-	protected String getJournalArticleResourceUuid(String journalArticleUuid)
-		throws Exception {
-
-		Connection con = null;
-		PreparedStatement ps = null;
-		ResultSet rs = null;
-
-		try {
-			con = DataAccess.getUpgradeOptimizedConnection();
-
-			StringBundler sb = new StringBundler(5);
-
-			sb.append("select JournalArticleResource.uuid_ from ");
-			sb.append("JournalArticleResource inner join JournalArticle on ");
-			sb.append("JournalArticle.resourcePrimKey = ");
-			sb.append("JournalArticleResource.resourcePrimKey where ");
-			sb.append("JournalArticle.uuid_ = ?");
-
-			ps = con.prepareStatement(sb.toString());
-
-			ps.setString(1, journalArticleUuid);
-
-			rs = ps.executeQuery();
-
-			if (rs.next()) {
-				return rs.getString("uuid_");
-			}
-
-			return null;
-		}
-		finally {
-			DataAccess.cleanUp(con, ps, rs);
-		}
 	}
 
 	@Override
@@ -224,12 +184,6 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 			portletPreferences.setValues("asset-entry-xml", assetEntryXmls);
 		}
 
-		if (ArrayUtil.isNotEmpty(assetEntryXmls)) {
-			upgradeUuids(assetEntryXmls);
-
-			portletPreferences.setValues("assetEntryXml", assetEntryXmls);
-		}
-
 		return PortletPreferencesFactoryUtil.toXML(portletPreferences);
 	}
 
@@ -260,36 +214,6 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 			assetEntryTypeElement.addText(assetEntryType);
 
 			rootElement.remove(assetTypeElement);
-		}
-	}
-
-	protected void upgradeUuids(String[] assetEntryXmls) throws Exception {
-		for (int i = 0; i < assetEntryXmls.length; i++) {
-			String assetEntry = assetEntryXmls[i];
-
-			Document document = SAXReaderUtil.read(assetEntry);
-
-			Element rootElement = document.getRootElement();
-
-			Element assetTypeElementUuid = rootElement.element(
-				"asset-entry-uuid");
-
-			String journalArticleResourceUuid = getJournalArticleResourceUuid(
-				assetTypeElementUuid.getStringValue());
-
-			if (journalArticleResourceUuid == null) {
-				continue;
-			}
-
-			rootElement.remove(assetTypeElementUuid);
-
-			assetTypeElementUuid.setText(journalArticleResourceUuid);
-
-			rootElement.add(assetTypeElementUuid);
-
-			document.setRootElement(rootElement);
-
-			assetEntryXmls[i] = document.formattedString(StringPool.BLANK);
 		}
 	}
