+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_0/UpgradeAssetPublisher.java
@@ -1,0 +1,131 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.upgrade.v7_0_0;
+
+import com.liferay.portal.kernel.dao.jdbc.DataAccess;
+import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
+import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.xml.Document;
+import com.liferay.portal.kernel.xml.Element;
+import com.liferay.portal.kernel.xml.SAXReaderUtil;
+import com.liferay.portlet.PortletPreferencesFactoryUtil;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+
+import javax.portlet.PortletPreferences;
+
+/**
+ * @author Sam Ziemer
+ */
+
+public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
+
+	protected String getJournalArticleResourceUuid(String journalArticleUuid)
+		throws Exception {
+
+		Connection con = null;
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+
+		try {
+			con = DataAccess.getUpgradeOptimizedConnection();
+
+			StringBundler sb = new StringBundler(5);
+
+			sb.append("select JournalArticleResource.uuid_ from ");
+			sb.append("JournalArticleResource inner join JournalArticle on ");
+			sb.append("JournalArticle.resourcePrimKey = ");
+			sb.append("JournalArticleResource.resourcePrimKey where ");
+			sb.append("JournalArticle.uuid_ = ?");
+
+			ps = con.prepareStatement(sb.toString());
+
+			ps.setString(1, journalArticleUuid);
+
+			rs = ps.executeQuery();
+
+			if (rs.next()) {
+				return rs.getString("uuid_");
+			}
+
+			return null;
+		}
+		finally {
+			DataAccess.cleanUp(con, ps, rs);
+		}
+	}
+
+	@Override
+	protected String[] getPortletIds() {
+		return new String[] {"101_INSTANCE_%"};
+	}
+
+	@Override
+	protected String upgradePreferences(
+			long companyId, long ownerId, int ownerType, long plid,
+			String portletId, String xml)
+		throws Exception {
+
+		PortletPreferences portletPreferences =
+			PortletPreferencesFactoryUtil.fromXML(
+				companyId, ownerId, ownerType, plid, portletId, xml);
+
+		String[] assetEntryXmls = portletPreferences.getValues(
+			"asset-entry-xml", new String[0]);
+
+		if (ArrayUtil.isNotEmpty(assetEntryXmls)) {
+			upgradeUuids(assetEntryXmls);
+
+			portletPreferences.setValues("assetEntryXml", assetEntryXmls);
+		}
+
+		return PortletPreferencesFactoryUtil.toXML(portletPreferences);
+	}
+
+	protected void upgradeUuids(String[] assetEntryXmls) throws Exception {
+		for (int i = 0; i < assetEntryXmls.length; i++) {
+			String assetEntry = assetEntryXmls[i];
+
+			Document document = SAXReaderUtil.read(assetEntry);
+
+			Element rootElement = document.getRootElement();
+
+			Element assetTypeElementUuid = rootElement.element(
+				"asset-entry-uuid");
+
+			String journalArticleResourceUuid = getJournalArticleResourceUuid(
+				assetTypeElementUuid.getStringValue());
+
+			if (journalArticleResourceUuid == null) {
+				continue;
+			}
+
+			rootElement.remove(assetTypeElementUuid);
+
+			assetTypeElementUuid.setText(journalArticleResourceUuid);
+
+			rootElement.add(assetTypeElementUuid);
+
+			document.setRootElement(rootElement);
+
+			assetEntryXmls[i] = document.formattedString(StringPool.BLANK);
+		}
+	}
+
+}
